feat: simplify build & usage
