+++ packages/catppuccin-vscode/.gitignore
@@ -1,2 +1,1 @@
 dist/
-themes/
