+++ packages/catppuccin-vscode/src/index.ts
@@ -1,11 +1,6 @@
-import { FlavorName } from "@catppuccin/palette";
-import { compileTheme, defaultOptions } from "catppuccin-vsc/src/theme";
-import { ThemeOptions } from "catppuccin-vsc/src/types";
+export { default as latte } from "catppuccin-vsc/themes/latte.json" with { type: "json" };
+export { default as frappe } from "catppuccin-vsc/themes/frappe.json" with { type: "json" };
+export { default as macchiato } from "catppuccin-vsc/themes/macchiato.json" with { type: "json" };
+export { default as mocha } from "catppuccin-vsc/themes/mocha.json" with { type: "json" };
 
-export const compile = (
-  flavor: FlavorName,
-  overrides: Partial<ThemeOptions> = {},
-) => {
-  const options = { ...defaultOptions, ...overrides };
-  return compileTheme(flavor, options);
-};
+export { compile } from "./compile.js";
