+++ packages/catppuccin-vscode/tsconfig.json
@@ -1,8 +1,8 @@
 {
   "extends": "@tsconfig/node20/tsconfig.json",
   "compilerOptions": {
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext",
+    "module": "esnext",
+    "moduleResolution": "Node",
     "noEmit": true,
     "paths": {
       "@/*": ["../catppuccin-vsc/src/*"]
